Split encode tests into internal and external tests to avoid import cycles
